+++ src/uint/rand.rs
@@ -1,7 +1,7 @@
 //! Random number generator support
 
 use super::Uint;
-use crate::{Limb, NonZero, Random, RandomMod};
+use crate::{Limb, NonZero, Random, RandomMod, RandomPrime};
 use rand_core::CryptoRngCore;
 use subtle::ConstantTimeLess;
 
@@ -50,9 +50,31 @@ impl<const LIMBS: usize> RandomMod for Uint<LIMBS> {
     }
 }
 
+impl<const LIMBS: usize> RandomPrime for Uint<LIMBS> {
+    /// Generate a cryptographically secure random [`Uint`].
+    ///
+    /// This function uses Miller-Rabin primality test with k=50 rounds
+    /// so the probability of returning a composite number is 4^(-50).
+    /// Note this method doesn't check whether the generated number
+    /// is a safe prime (i.e. that (p-1)/2 is also prime).
+    ///
+    /// The variable-time nature of the algorithm should not pose a security
+    /// issue so long as the underlying random number generator is truly a
+    /// CSRNG, where previous outputs are unrelated to subsequent
+    /// outputs and do not reveal information about the RNG's internal state.
+    fn random_prime(rng: &mut impl CryptoRngCore) -> Self {
+        loop {
+            let candidate = Self::random(rng);
+            if candidate.is_prime_miller_rabin_vartime(rng, 50) {
+                return candidate;
+            }
+        }
+    }
+}
+
 #[cfg(test)]
 mod tests {
-    use crate::{NonZero, RandomMod, U256};
+    use crate::{NonZero, RandomMod, RandomPrime, U1024, U256};
     use rand_core::SeedableRng;
 
     #[test]
@@ -75,5 +97,18 @@ mod tests {
         // Check that the value is in range
         assert!(res >= U256::ZERO);
         assert!(res < U256::from(0x10000000000000001u128));
+    }
+
+    #[test]
+    fn random_prime() {
+        let mut rng = rand_chacha::ChaCha8Rng::seed_from_u64(1);
+
+        let res = U1024::random_prime(&mut rng);
+        // openssl prime -hex DC1DAB4D998BEA51033B81FA7CB97D92B19C11313F04C663D9E8075C2AD0F847AB1E4B28D0E0A6C58A87658B4512FB693C10DEC5341895D855A26C98A259C74B8AC1D7F1F57B56B9D6FB99D23BFAC23B21EE44872E9E9C9BBD22C4301D9BF1B5C80C0A33C139F0BE4477A4EED81ABCE305AD35FE5D598BE2C6FE0EC445F6EDA5
+        // outputs "is prime"
+        assert_eq!(
+            U1024::from_be_hex("DC1DAB4D998BEA51033B81FA7CB97D92B19C11313F04C663D9E8075C2AD0F847AB1E4B28D0E0A6C58A87658B4512FB693C10DEC5341895D855A26C98A259C74B8AC1D7F1F57B56B9D6FB99D23BFAC23B21EE44872E9E9C9BBD22C4301D9BF1B5C80C0A33C139F0BE4477A4EED81ABCE305AD35FE5D598BE2C6FE0EC445F6EDA5"), 
+            res
+        );
     }
 }